renamed to locales
